Solve directory navigation problem by adding '/' to the end of directory links
Currently, it seems that browsers directory is not handled correctly when a link to directory does
not have '/' at the end. The problem seems to be in the browser side. Links are not relatively correct.
When browser calls IP/my_documents, the current directory in browser side is not changed to my_documents,
so links like '../' will not work!. So if navigation is needed, should give a url with '/' at the end.